+++ plugins/woocommerce-admin/packages/experimental/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 -   Remove the use of Dashicons and replace with @wordpress/icons or gridicons #7020
 -   Add expanded item text and CTA button. #6956
+-   Add inbox note components (InboxNoteCard, InboxNotePlaceholder, and InboxDismissConfirmationModal). #7006
 
 # 1.2.0